// types.ts
export interface TeamProjectPayment {
  id: string;
  projectId: string;
  teamMemberName: string;
  teamMemberId: string;
  date: string;
  status: 'Paid' | 'Unpaid';
  fee: number;
  reward?: number;
}